lib: Split launch.c into separate files.
launch-appliance.c contains the code associated with the 'appliance'
attach-method.  Mostly.  In fact there are a few APIs which don't fit
so nicely:

 - config: deprecated API which fiddles with the qemu command
   line directly

 - max-disks: depends on the qemu implementation (virtio-scsi
   or not)

 - debug-drives: used for testing only

launch-unix.c contains the code associated with 'unix:<path>'.

launch.c is the common code for launching, along with a few other APIs
such as guestfs_add_drive_opts.

This commit also reduces the number of headers to just those
which are required.