sparsify: Use automake to build and link virt-sparsify binary.
There are two observations which make this possible: (1) virt-sparsify
contains at least one C unit, so automake thinks it is compiling a C
program.  (2) Automake lets us add a *_LINK rule to override linking.

We list the C sources of virt-sparsify as virt_sparsify_SOURCES, so
automake thinks it is building a C program as normal.

We override virt_sparsify_LINK to make it actually link an OCaml
program.

We add virt_sparsify_DEPENDENCIES listing the OCaml objects so that
they get built before linking.

We need a small linker script (link.sh) which adds some options at the
end of the linker command line which are impossible to add using pure
automake.

Inspired by the same idea in supermin:
https://github.com/libguestfs/supermin/blob/master/src/Makefile.am