builder: Fix parallel builds of index-parse.o.
When parallel builds are enabled it was possible for index-parse.c to
be compiled before the file had been completely written by bison.  The
usual symptom was that the 'do_parse' symbol was missing -- this
simply happened because that symbol is defined at the end of this
file, and the compiler compiled the file before the end part was
written out.  But you could also see other strange & random compile
failures as you would expect.

This was tested by running this script on an 8 core server:

 cd builder/
 export MAKEFLAGS=-j9
 while make clean >/dev/null; make all V=1 >& /tmp/log; do echo -n . ; done

Previously it would fail after <= 5 iterations.  With this change it
runs for at least 100 iterations.

(cherry picked from commit df5bd5741b37da9cf97d7a76ac2805557aa630db)